refactor: fix comment component
