+++ src/index.js
@@ -23,5 +23,6 @@ reportWebVitals();
 /*
 
   - Loader не работает при открытии новости
+  - Если нет интернета, то все падает
 
 */
